+++ packages/widgetbook/test/src/addons/utils/cupertino_app_theme.dart
@@ -24,7 +24,7 @@ final cupertinoThemeSetting = CupertinoThemeSetting(
 );
 
 final cupertinoThemeAddon = CupertinoThemeAddon(
-  themeSetting: cupertinoThemeSetting,
+  setting: cupertinoThemeSetting,
 );
 
 AppBuilderFunction get cupertinoAppBuilder =>