+++ packages/widgetbook_cli/bin/models/models.dart
@@ -1,5 +1,5 @@
 export './ci_args.dart';
 export './cli_args.dart';
-export './deployment_data.dart';
+export './create_build_request.dart';
 export './package.dart';
 export './review_data.dart';