+++ packages/widgetbook_cli/bin/models/review_data.freezed.dart
@@ -20,11 +20,6 @@ mixin _$ReviewData {
   String get buildId => throw _privateConstructorUsedError;
   String get projectId => throw _privateConstructorUsedError;
   String get baseSha => throw _privateConstructorUsedError;
-  List<ThemeData> get themes => throw _privateConstructorUsedError;
-  List<LocaleData> get locales => throw _privateConstructorUsedError;
-  List<DeviceData> get devices => throw _privateConstructorUsedError;
-  List<TextScaleFactorData> get textScaleFactors =>
-      throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
   $ReviewDataCopyWith<ReviewData> get copyWith =>
@@ -40,11 +35,7 @@ abstract class $ReviewDataCopyWith<$Res> {
       {List<ChangedUseCase> useCases,
       String buildId,
       String projectId,
-      String baseSha,
-      List<ThemeData> themes,
-      List<LocaleData> locales,
-      List<DeviceData> devices,
-      List<TextScaleFactorData> textScaleFactors});
+      String baseSha});
 }
 
 /// @nodoc
@@ -61,10 +52,6 @@ class _$ReviewDataCopyWithImpl<$Res> implements $ReviewDataCopyWith<$Res> {
     Object? buildId = freezed,
     Object? projectId = freezed,
     Object? baseSha = freezed,
-    Object? themes = freezed,
-    Object? locales = freezed,
-    Object? devices = freezed,
-    Object? textScaleFactors = freezed,
   }) {
     return _then(_value.copyWith(
       useCases: useCases == freezed
@@ -83,22 +70,6 @@ class _$ReviewDataCopyWithImpl<$Res> implements $ReviewDataCopyWith<$Res> {
           ? _value.baseSha
           : baseSha // ignore: cast_nullable_to_non_nullable
               as String,
-      themes: themes == freezed
-          ? _value.themes
-          : themes // ignore: cast_nullable_to_non_nullable
-              as List<ThemeData>,
-      locales: locales == freezed
-          ? _value.locales
-          : locales // ignore: cast_nullable_to_non_nullable
-              as List<LocaleData>,
-      devices: devices == freezed
-          ? _value.devices
-          : devices // ignore: cast_nullable_to_non_nullable
-              as List<DeviceData>,
-      textScaleFactors: textScaleFactors == freezed
-          ? _value.textScaleFactors
-          : textScaleFactors // ignore: cast_nullable_to_non_nullable
-              as List<TextScaleFactorData>,
     ));
   }
 }
@@ -114,11 +85,7 @@ abstract class _$$_ReviewDataCopyWith<$Res>
       {List<ChangedUseCase> useCases,
       String buildId,
       String projectId,
-      String baseSha,
-      List<ThemeData> themes,
-      List<LocaleData> locales,
-      List<DeviceData> devices,
-      List<TextScaleFactorData> textScaleFactors});
+      String baseSha});
 }
 
 /// @nodoc
@@ -137,10 +104,6 @@ class __$$_ReviewDataCopyWithImpl<$Res> extends _$ReviewDataCopyWithImpl<$Res>
     Object? buildId = freezed,
     Object? projectId = freezed,
     Object? baseSha = freezed,
-    Object? themes = freezed,
-    Object? locales = freezed,
-    Object? devices = freezed,
-    Object? textScaleFactors = freezed,
   }) {
     return _then(_$_ReviewData(
       useCases: useCases == freezed
@@ -159,22 +122,6 @@ class __$$_ReviewDataCopyWithImpl<$Res> extends _$ReviewDataCopyWithImpl<$Res>
           ? _value.baseSha
           : baseSha // ignore: cast_nullable_to_non_nullable
               as String,
-      themes: themes == freezed
-          ? _value._themes
-          : themes // ignore: cast_nullable_to_non_nullable
-              as List<ThemeData>,
-      locales: locales == freezed
-          ? _value._locales
-          : locales // ignore: cast_nullable_to_non_nullable
-              as List<LocaleData>,
-      devices: devices == freezed
-          ? _value._devices
-          : devices // ignore: cast_nullable_to_non_nullable
-              as List<DeviceData>,
-      textScaleFactors: textScaleFactors == freezed
-          ? _value._textScaleFactors
-          : textScaleFactors // ignore: cast_nullable_to_non_nullable
-              as List<TextScaleFactorData>,
     ));
   }
 }
@@ -186,16 +133,8 @@ class _$_ReviewData implements _ReviewData {
       {required final List<ChangedUseCase> useCases,
       required this.buildId,
       required this.projectId,
-      required this.baseSha,
-      required final List<ThemeData> themes,
-      required final List<LocaleData> locales,
-      required final List<DeviceData> devices,
-      required final List<TextScaleFactorData> textScaleFactors})
-      : _useCases = useCases,
-        _themes = themes,
-        _locales = locales,
-        _devices = devices,
-        _textScaleFactors = textScaleFactors;
+      required this.baseSha})
+      : _useCases = useCases;
 
   final List<ChangedUseCase> _useCases;
   @override
@@ -210,37 +149,10 @@ class _$_ReviewData implements _ReviewData {
   final String projectId;
   @override
   final String baseSha;
-  final List<ThemeData> _themes;
-  @override
-  List<ThemeData> get themes {
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_themes);
-  }
-
-  final List<LocaleData> _locales;
-  @override
-  List<LocaleData> get locales {
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_locales);
-  }
-
-  final List<DeviceData> _devices;
-  @override
-  List<DeviceData> get devices {
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_devices);
-  }
-
-  final List<TextScaleFactorData> _textScaleFactors;
-  @override
-  List<TextScaleFactorData> get textScaleFactors {
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_textScaleFactors);
-  }
 
   @override
   String toString() {
-    return 'ReviewData(useCases: $useCases, buildId: $buildId, projectId: $projectId, baseSha: $baseSha, themes: $themes, locales: $locales, devices: $devices, textScaleFactors: $textScaleFactors)';
+    return 'ReviewData(useCases: $useCases, buildId: $buildId, projectId: $projectId, baseSha: $baseSha)';
   }
 
   @override
@@ -251,12 +163,7 @@ class _$_ReviewData implements _ReviewData {
             const DeepCollectionEquality().equals(other._useCases, _useCases) &&
             const DeepCollectionEquality().equals(other.buildId, buildId) &&
             const DeepCollectionEquality().equals(other.projectId, projectId) &&
-            const DeepCollectionEquality().equals(other.baseSha, baseSha) &&
-            const DeepCollectionEquality().equals(other._themes, _themes) &&
-            const DeepCollectionEquality().equals(other._locales, _locales) &&
-            const DeepCollectionEquality().equals(other._devices, _devices) &&
-            const DeepCollectionEquality()
-                .equals(other._textScaleFactors, _textScaleFactors));
+            const DeepCollectionEquality().equals(other.baseSha, baseSha));
   }
 
   @override
@@ -265,11 +172,7 @@ class _$_ReviewData implements _ReviewData {
       const DeepCollectionEquality().hash(_useCases),
       const DeepCollectionEquality().hash(buildId),
       const DeepCollectionEquality().hash(projectId),
-      const DeepCollectionEquality().hash(baseSha),
-      const DeepCollectionEquality().hash(_themes),
-      const DeepCollectionEquality().hash(_locales),
-      const DeepCollectionEquality().hash(_devices),
-      const DeepCollectionEquality().hash(_textScaleFactors));
+      const DeepCollectionEquality().hash(baseSha));
 
   @JsonKey(ignore: true)
   @override
@@ -279,15 +182,10 @@ class _$_ReviewData implements _ReviewData {
 
 abstract class _ReviewData implements ReviewData {
   factory _ReviewData(
-          {required final List<ChangedUseCase> useCases,
-          required final String buildId,
-          required final String projectId,
-          required final String baseSha,
-          required final List<ThemeData> themes,
-          required final List<LocaleData> locales,
-          required final List<DeviceData> devices,
-          required final List<TextScaleFactorData> textScaleFactors}) =
-      _$_ReviewData;
+      {required final List<ChangedUseCase> useCases,
+      required final String buildId,
+      required final String projectId,
+      required final String baseSha}) = _$_ReviewData;
 
   @override
   List<ChangedUseCase> get useCases => throw _privateConstructorUsedError;
@@ -297,15 +195,6 @@ abstract class _ReviewData implements ReviewData {
   String get projectId => throw _privateConstructorUsedError;
   @override
   String get baseSha => throw _privateConstructorUsedError;
-  @override
-  List<ThemeData> get themes => throw _privateConstructorUsedError;
-  @override
-  List<LocaleData> get locales => throw _privateConstructorUsedError;
-  @override
-  List<DeviceData> get devices => throw _privateConstructorUsedError;
-  @override
-  List<TextScaleFactorData> get textScaleFactors =>
-      throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
   _$$_ReviewDataCopyWith<_$_ReviewData> get copyWith =>