+++ packages/widgetbook_generator/lib/models/widgetbook_use_case_data.g.dart
@@ -19,6 +19,7 @@ _$_WidgetbookUseCaseData _$$_WidgetbookUseCaseDataFromJson(
           .toList(),
       componentDefinitionPath: json['componentDefinitionPath'] as String,
       useCaseDefinitionPath: json['useCaseDefinitionPath'] as String,
+      designLink: json['designLink'] as String?,
     );
 
 Map<String, dynamic> _$$_WidgetbookUseCaseDataToJson(
@@ -32,4 +33,5 @@ Map<String, dynamic> _$$_WidgetbookUseCaseDataToJson(
       'dependencies': instance.dependencies,
       'componentDefinitionPath': instance.componentDefinitionPath,
       'useCaseDefinitionPath': instance.useCaseDefinitionPath,
+      'designLink': instance.designLink,
     };