+++ packages/widgetbook_generator/test/src/code_generators/instances/addons/frame_addon/frames/widgetbook_frame/widgetbook_frame_instance_test.dart
@@ -1,0 +1,19 @@
+import 'package:test/test.dart';
+import 'package:widgetbook_generator/code_generators/instances/addons/frame_addon/frames/frames.dart';
+import 'package:widgetbook_models/widgetbook_models.dart';
+
+void main() {
+  group(
+    '$WidgetbookFrameInstance',
+    () {
+      test('.name is "WidgetbookFrame"', () {
+        final sut = WidgetbookFrameInstance(
+          devices: const [
+            Apple.iPhone11,
+          ],
+        );
+        expect(sut.name, equals('WidgetbookFrame'));
+      });
+    },
+  );
+}